requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
docutils==0.17.1
flex==6.14.1
genson==1.2.2
idna==2.10
importlib-metadata==4.5.0
jsonpath-ng==1.5.2
jsonpickle==2.0.0
jsonpointer==2.1
jsonschema==3.2.0
ply==3.11
Pygments==2.9.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
RESTinstance==1.1.0
rfc3987==1.3.8
robotframework==4.0.3
robotframework-robocop==1.7.1
robotframework-tidy==1.1.1
six==1.16.0
strict-rfc3339==0.7
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
validate-email==1.3
zipp==3.4.1
